+++ R/sAttributes2cpos_method.R
@@ -1,23 +1,26 @@
 setGeneric("sAttributes2cpos", function(.Object, ...) standardGeneric("sAttributes2cpos"))
 
-#' Get corpus positions based on s-attributes.
+#' Get corpus position matrix.
 #' 
 #' Based on the sAttributes as defined in the respective slot of the partition
-#' object, corpus positions (matrix in the cpos slot) is generated.
+#' object, a corpus position matrix is generated, i.e. the slot \code{cpos}
+#' of the object is filled.
 #' 
-#' @param partition the rudimentary partition object
+#' @param partition partition object with list of s-attributes
 #' @param xml either "flat" or "nested"
 #' @param regex logical
 #' @return An augmented partition object, with cpos and strucs slots filled.
 #' @noRd
-setMethod("sAttributes2cpos", "partition", function(.Object, xml, regex){
+setMethod("sAttributes2cpos", "partition", function(.Object, xml = "flat", regex = FALSE){
+  stopifnot(xml %in% c("flat", "nested"))
+  stopifnot(is.logical(regex))
   if (xml == "flat"){
     # The function works nicely - potentially, it can be optimized, but I have tried many things.
     # Interestingly, the for-loop is more effective than a vectorized version
     # an Rcpp-implementation of struc2str is not faster
     # potential for optimization: struc2str
     maxAttr <- CQI$attribute_size(.Object@corpus, .Object@sAttributeStrucs, type = "s")
-    meta <- data.frame(struc = 0:(maxAttr-1), select = rep(0, times = maxAttr))
+    meta <- data.frame(struc = 0:(maxAttr - 1), select = rep(0, times = maxAttr))
     if (length(.Object@sAttributes) > 0) {
       for (sAttr in names(.Object@sAttributes)){
         meta[,2] <- as.vector(CQI$struc2str(.Object@corpus, sAttr, meta[,1]))
@@ -77,9 +80,9 @@ setMethod("sAttributes2cpos", "partition", function(.Object, xml, regex){
         sAttrValues <- CQI$struc2str(.Object@corpus, sAttrNames[i], CQI$cpos2struc(.Object@corpus, sAttrNames[i], cpos[,1]))
         Encoding(sAttrValues) <- .Object@encoding
         if (regex == FALSE) {
-          hits <- which(meta %in% .Object@sAttributes[[ sAttrNames[i] ]])
+          hits <- which(sAttrValues %in% .Object@sAttributes[[ sAttrNames[i] ]])
         } else if (regex == TRUE) {
-          hits <- unique(unlist(lapply(.Object@sAttributes[[ sAttrNames[i] ]], function(x) grep(x, meta))))
+          hits <- unique(unlist(lapply(.Object@sAttributes[[ sAttrNames[i] ]], function(x) grep(x, sAttrValues))))
         }
         cpos <- cpos[hits,]
         strucs <- strucs[hits]